Fix -Waddress warnings in nptl/tst-mutex1.c.
This patch fixes -Waddress warnings in nptl/tst-mutex1.c from
comparing the address of an object with NULL (ATTR may either be NULL,
or the address of an object when included from other tests, and the
warning arises in the latter case).  A macro ATTR_NULL is defined
alongside ATTR and used for the tests.

Tested for x86_64.

	* nptl/tst-mutex1.c: Include <stdbool.h>.
	[!ATTR] (ATTR_NULL): New macro.
	(do_test): Test !ATTR_NULL instead of ATTR != NULL.
	* nptl/tst-mutexpi1.c (ATTR_NULL): New macro.
	* nptl/tst-mutexpp1.c (ATTR_NULL): New macro.