Disable -Wdeprecated-declarations for register_printf_function calls in tst-printfsz.c.
This patch uses diagnostic control pragmas to disable warnings in
stdio-common/tst-printfsz.c for use of the deprecated
register_printf_function.  Because this test is testing printf_size
and printf_size_info, and the latter has the interface expected for
register_printf_function instead of the newer
register_printf_specifier, it seems correct for this test to use the
deprecated interface (wrapping printf_size_info in some way to use
register_printf_specifier would seem an excessive change to what's
tested).

Tested for x86_64.

	* stdio-common/tst-printfsz.c: Include <libc-internal.h>.
	(main): Disable -Wdeprecated-declarations around calls to
	register_printf_function.