Fix MIPS _ABIO64 -Werror=undef build.
I see an error

../sysdeps/mips/memcpy.S:209:68: error: "_ABIO64" is not defined [-Werror=undef]
 #if defined(_MIPS_SIM) && ((_MIPS_SIM == _ABIO32) || (_MIPS_SIM == _ABIO64))
                                                                    ^
cc1: some warnings being treated as errors

in MIPS builds.  This patch arranges for _ABIO64 to be defined with
the same value as GCC uses when building for O64 (the ABI itself isn't
supported by glibc, but defining the macro seems the simplest way of
avoiding the error in code that may be shared with other C libraries).

	* sysdeps/mips/sgidefs.h [!_ABIO64] (_ABIO64): New macro.